+++ lib/l10n/app_localizations_sw.dart
@@ -123,6 +123,9 @@ class AppLocalizationsSw extends AppLocalizations {
   String get licensesButton => 'Leseni za chanzo huria';
 
   @override
+  String get privacyPolicyButton => 'Sera ya faragha';
+
+  @override
   String get sourceCodeButton => 'Msimbo wa chanzo kwenye GitHub';
 
   @override